Bootcamp ownership

+++ controllers/bootcamps.js
@@ -53,10 +53,7 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route   PUT /api/v1/bootcamps/:id
 // @access  Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
       return next(
@@ -64,6 +61,16 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
       );
     }
 
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+      );
+    }
+    bootcamp = await Bootcamp.findOneAndUpdate({"_id": req.params.id }, req.body, {
+      new: true,
+      runValidators: true
+    })
     res.status(200).json({ success: true, data: bootcamp });
 });
 
@@ -71,14 +78,21 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 // @route   DELETE /api/v1/bootcamps/:id
 // @access  Private
 exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndDelete(req.params.id);
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
       );
     }
-
+  
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`, 401)
+      );
+    }
+    bootcamp = await Bootcamp.findByIdAndDelete(req.params.id);
     await bootcamp.deleteOne();
 
     res.status(200).json({ success: true, data: {} });
@@ -123,6 +137,14 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
+
+  // Make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+    );
+  }
+
   if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 400));
   }
